Update Raspbian.shtml
This is page for installing JMRI on a Raspberry Pi . This is also a separate hardware page in the JMRI repository related to using Raspberry Pi's GPIO pins. I have put in links to cross-reference so that naive user (e.g. me) will realize that these are different and both are possible.